Update manpage mentions of for_signature
In the Builder Objects section, under generator, just add a note about
omitting pieces that don't affect rebuilds if for_signature is true.

In the Variable Substitution section, the same is added for a
variable referring to a function, but a little mroe wording added,
and since the new info kind of depends on the $( $) construct,
which came just below it in the section, that chunk is moved up to
make the reference a little more natural.

Signed-off-by: Mats Wichmann <mats@linux.com>