Update docs for HOST_OS, TEMPFILE and Split  [skip travis] [skip appveyor]
The HOST_ARCH/TARGET_ARCH functions have been defined in two
places, generic and Windows, these are consolidated into one,
leaving a comment in the msvc.xml file to show where that version
went to.  HOST_OS/TARGET_OS also got some tweaks - this was in
line with these now being set by default for all platforms.

The $TERMPFILE construction var was not defined; added and the
whole family had docs updated.

A small tweak to Split wording.

References to Python changed to entity references (there are still
more of these to change).

lambda as an entity reference is not correct (generated docs show
the symbol, but Python does not use the lambda symbol to refer
to lambda expressions).

Signed-off-by: Mats Wichmann <mats@linux.com>