Skip unusable javah tests
For the three tests which fail, skip if the command finder (WhereIs), which
is much less aggressive than the java_where_j* routines, doesn't find javah.

Split the JAVAH test file into two files, so that the part which
doesn't call javah can run even if the rest is skipped.

Did the header cleanup and partial reformats on those.

Leave a comment in the Repository/JavaH test, and do some reformatting
there as well.  Test needs rework, but for now doesn't fail so leave alone.

Signed-off-by: Mats Wichmann <mats@linux.com>