Add an extra check for javah
The previous sanity check was to see if the less-optimistic
WhereIs could find javah, and skip if not, since java_where_javah
looks many more places than SCons will actually look.  This
was insufficient - some openjdk packages don't clean up the
alternatives link to javah even if a java version which doesn't
have it becomes the selected version. Add a check on found
paths to detect a mismatch.

Signed-off-by: Mats Wichmann <mats@linux.com>