Skip two scons-time tests on Windows
The scons-time tests trying to test the "run" subcommand of
scons-time substitute a dummy script named my_python.py for Python.
In order for this to work on Windows, the file association chain for
.py files has to lead back to an Python interpreter, but we cannot
control this - for example, systems with VS Code installed often
steals this association in ways that are aren't easy to pre-detect.

Note that scons-time itself carfully uses the Python interpreter
to launch things so it's not affected by this, it's just the test
that tries to mock part of this scheme that has problems.

Signed-off-by: Mats Wichmann <mats@linux.com>