Followon test fix for PR #4067
There was another hole in the tests after updating them, which was
missed locally because the local test machines took the "skip for
external reasons" path, and didn't actually run the test that failed.
*Any* SCons test path starts with the value of %TEMP%, which by default
is "C:\Users\something\AppData\Local\Temp".  If not entered as a raw
string, that's a unicode escape starting with the 3rd char, and leads to
a decode error.  I somehow uncovovered one of those that wasn't protected
as a rawstring.

Signed-off-by: Mats Wichmann <mats@linux.com>