Update manpage "Extending SCons" section.  [skip appveyor]
Now organized into four sections: Builder Objects, Action Objects,
Scanner Objects, Tool Modules, with the latter (new) section getting a
lot of the material that was in the Tools section (which users dodn't
need just to select an existing tool).

Adds a bit of commentary to the beginning of the section, and
to Builder and Scanner.

Signed-off-by: Mats Wichmann <mats@linux.com>