Fix rst2pdf/Sphinx problem a better way
Our Sphinx config specifies the use of rst2pdf:
extensions = [ ... 'rst2pdf.pdfbuilder', ...
As of 0.99 we apparently need to declare a pdf stylesheet, even though
we don't knowingly use that style.

With that change, unpin rst2pdf again. Also upgrade lxml, leaving
it pinned to a version with a comment why.

Signed-off-by: Mats Wichmann <mats@linux.com>