Massage test fixtures a bit
To avoid double loops, a couple now use fileinput module.

CC and CXX tests switch to using general mylink.py fixture instead of
locally defined linker script.  This did mean that the marker for the
mock linker that the script removes had to be harmonized as #link - some
were using /*link*/.

Could not switch CC/CXX to use the general mycompile.py mock compiler
due to difference in intent: mycompile.py is intended to plug in as a
*COM variable, which means the whole cmdline is in the caller's control.
mycc.py is intended to plug in as CC and mycxx.py as CXX, which means
standard os-dependent construction vars will be passed via the default
CCCOM and CXXCOM and have to be handled.

Signed-off-by: Mats Wichmann <mats@linux.com>