Tweak the swig tool documentation  [skip appveyor]
Minor changes: the spelled-out name of the tool as used in SCons docs
was not the same as the manpage uses, and it was repeated a bunch of
times spelled out for no good reason.  Just turned it all into uses of
&swig; entity.

A bit of tweaking of some of the entries.  SWIGCXXFILESUFFIX had a
copy-paste error, it referred to CFILESUFFIX instead of CXXFILESUFFIX.

Signed-off-by: Mats Wichmann <mats@linux.com>