Tweak the scons help message
Processing changed a bit - some lines can now be joined instead of
split (if the usage part is short so the combination will still fit).
Dropped the "Ingored for compatibility" chunk from the printout.

The usage: message was changed as it didn't mention variables.  A number
of tests expected the exact value of that line, and so were updated.

Updated docstrings (for the API docs).

Signed-off-by: Mats Wichmann <mats@linux.com>