PR 4111: tweak msvc cache invalid algorithm
Adapt previous change so if cache entry tools dir is empty, don't
invalidate the entry, it's an old compiler that doesn't add this info
(we always add it, if it didn't come from the batch file it will be an
empty list). We don't want to force that case to recalculate, or the
value of the cache is lost for those compilers.

Signed-off-by: Mats Wichmann <mats@linux.com>