Fix content-timestamp decider for symlinks
The base filesystem node class has getmtime() and getsize() functions.
Those were changed in an early commit to use the lstat() method if
the node represented a symbolic link. However, we actually want the
information of the file the symlink points to, or we can't detect
changes to the mtime or size of the underlying file, and miss rebuilds
if content-timestamp is used.

Added a testcase which shows the failure to rebuild from the symlinked
source.

Fixes #3880

Signed-off-by: Mats Wichmann <mats@linux.com>