Revert to using stat.ST_MTIME in FS nodes for now
Part of the change to restore mtime checking of symlinked sources
flipped from statobj[stat.ST_MTIME] to statobj.st_mtime. These return
"the same thing" but the latter returns a float instead of an int.
This had some unexpected side effects - like failing a half-dozen
testcases.  In the interest of moving the bugfix along, revert this
piece and deal with it some other time.

Signed-off-by: Mats Wichmann <mats@linux.com>