test framework: fix exception on timeout
If the framework wait_for() method actually times out, it tries to return
stdout and stderr by calling the framework methods of those names.
The stdout() method was protected against the message not having been
captured (as is the case on timeout).  Updated the stderr() method to
use the same technique.

Signed-off-by: Mats Wichmann <mats@linux.com>