Fortran vairants now include FORTRANCOMMONFLAGS
The documentation suggested $FORTRANFLAGS was included in the build
lines for all variants, but it was not.  Turns out the test suite
quite explicitly checks that FORTRANFLAGS doesn't leak through to other
variants, so instead define a new FORTRANCOMMONFLAGS to serve the purpose
of a flag that applies to all variants.

Assorted cleanup. And f-strings.

Fixes #2257

Signed-off-by: Mats Wichmann <mats@linux.com>