find_prorgram_path() can now add discovered path
Many tools contain a stanza like this, usually for the Windows case,
where installed programs are rarely added to the "standard" path
in SCons' execution environment (i.e. env['ENV']["PATH']):

    javac = SCons.Tool.find_program_path(env, 'javac', default_paths=paths)
    if javac:
        javac_bin_dir = os.path.dirname(javac)
        env.AppendENVPath('PATH', javac_bin_dir)

This change adds a keyword argument add_path to find_program_path() to
instruct it to add the path a program was discovered in, if it was in the
extra paths passed in. The default is False, retaining the existing behavior.

Signed-off-by: Mats Wichmann <mats@linux.com>