Improvements to lex and yacc tools
The mocked tools mylex.py and myyacc.py now understand the file-generation
options, and generate a dummy file with predictable contents, for
checking.  This allows more testing of the path through the SCons support
for these two without needing live commands.

New tests added which invoke the file-generation options, and make
sure the extra files are created, and that SCons detects and tracks
the added targets.  Work is done in a subdirectory, which exposes some
existing known inconsistent behavior (the regular generated file goes
in the subdir per the LEXCOM and YACCOM generated line, while the ones
generated from commandline options go in the topdir) - but we're going
to allow that behavior to continue for backwards compat.

Same fix applied to yacc tool that PR #4168 did for lex - do subst_list()
instead of subst() to preserve spaces in paths.  That fix left the lex
tool unable to pass the new test, as it could not see the individual
arguments in the FLAGS variable, which was solved by indexing into the
subst'd list so we can iterate over the args again.

Test and tool cleanup; add DefaultEnvironment calls, etc.

Note this mentions, but does not address the problem described in issue 4154.

Signed-off-by: Mats Wichmann <mats@linux.com>