Ajust test for 3.11+ enhanced error messages
There were four tests failing because the regexes for expected
stdout didn't match with the addition of pointers in exceptions showing
where the error occurs.  However, with 3.11.0b4, this has been backed
off a bit: there are no arrow lines if the whole line would have been
highlighted, since that adds no useful new information.  As a result,
only one test now failed, and the regex there is expanded to handle
that case.

This relates to, but does not close, issue #4162.

Signed-off-by: Mats Wichmann <mats@linux.com>