Stop using deprecated load_module
importlib.load_module is deprecated since 3.10 and scheduled for removal
in the 2023 Python release (3.12) - this makes the tests somewhat noisy.

The zip import code was reworked, to use exec_module.  A test was added
in the Platform code - there's now a dummy platform module inside a zip
file that the PlatformTests unittest uses.

The tool import code finally dropped the fallback to Python2-style
importing, and the zip import section matches the (now tested) Platform
version.  Not sure this is needed at all: the regular import machinery
ought to find a zipfile without extra steps, since here it uses sys.path
(the platform code does *not* use sys.path).

sconsign now just uses import_module, the "my_import" function is no
longer needed but was left in for now.  sconsign actually used the "imp"
module, which is deprecated since 3.4.

A little cleaup in the three scripts/ files.

Fixes #4162

Signed-off-by: Mats Wichmann <mats@linux.com>