load_module work: back off zipimport changes
Apparently, the zipimport code was not intended to load modules
(platform, tool) that were stored in individual zipfiles;
instead it was added to support scons itself running inside a
zipfile in the form of a py2exe bundle. Undid the changes which
tried to find a modulename.zip file and load it, and the test
in Platform that actually verfied it works, no need to add a
new feature nobody is asking for.

Signed-off-by: Mats Wichmann <mats@linux.com>