Dedent some examples in userguide  [ci skip]
The compilation DB examples in the User Guide are indented with
the rest of the text (possibly auto-formatting by an editor),
while the scons_example and programlisting XML elements respect
formatting, so they should actually be dedented.

Example code which is not to be displayed (just used to drive
the example generation) don't matter so they were left alone this time.

Signed-off-by: Mats Wichmann <mats@linux.com>