[framework] quiet Python about start() tests
Recent Pythons raise a ResourceWarning if they collect resources
that had not been closed.  The TestCmdTest tests have a testcase for
the "start" method that because of the way it was written, does
not close subprocess.Popen files before starting a new instance,
which leads to the collection/warnig. Add a _cleanup method in that
test class for closing.

Signed-off-by: Mats Wichmann <mats@linux.com>