Convert line endings on sources   [skip appveyor]
Some file still were checked in with DOS-style line endings. Convert
for consistency, and let git handle line endings on checkout for the
platform in question.

Did *not* convert svg files.

Added .git-blame-ignore-revs with this revision listed to be able
to ignore this change when doing "git blame --ignore-revs-file=...".

To not have to manually type the option, can do:
    git config --global blame.ignoreRevsFile .git-blame-ignore-revs

Signed-off-by: Mats Wichmann <mats@linux.com>