WIP: Update process spawning in platforms
Default is now simplified functions that just use subprocess.run().
On posix, the escape function is now shlex.quote from stdlib.  On win32,
some funky dances with temporary files are eliminated, as subprocess is
quite capable of capturing output.

In Subst, quote_spaces is generalized a bit, and a TODO is left in the
comments - it doesn't look like CmdStringHolder.escape ever gets called
with a quote_func argument, so the current way is a bit inefficient,
but it's unlikely to be a compelling improvement so seems not worth the
research to be sure "nothing breaks".

The change of escape function on posix means a couple of places in one
test see a change: some arguments that don't need quoting are no longer
quoted. Specifically, it's these two former expectations:

BF:    %(_python_)s myfail.py f03 f04 "f04" "f04.in"
BF:    %(_python_)s myfail.py f04 f05 "f05" "f05.in"

It's possible that the former slightly funky posix quoting, was to match
the win32 output, so tests like this can work with the same output on
both flavors.  The test is modified to paste in a platform-specific
variable to account for the current difference. Going back to the old
quoting style is, in a sense, disjoint from the change to use subprocess.

Signed-off-by: Mats Wichmann <mats@linux.com>