Handle a list as source in Copy
Copy (in the copy_func()) unconditionally converts the source argument
to a string.  This breaks the case where it was passed as a list -
it should remain a list with the members converted to string, rather than
being a single string. We assume a list source is supported, since the
code just a few lines lower down detects a list and iterates through it.

A test is added in test/Copy-Action.py to try to detect this case -
it fails before the change and passes after.

With the change to make a list argument work, found that the message
printed by the matching strfunction (originally a lambda) needed the
same adjustment, if the caller did not supply a list that was already
strings (e.g. a list of Nodes, as from calling Glob()). Fixed up.

At the same time, did some fiddling with the other Action Functions.
Chmod had not been adjusted to emit the new Python standard for octal
values (0o755 instead of 0755) - the change is only for the message,
not for the function, but required aligning the test with the new format.

Added some docstrings.

Fixes #3009

Signed-off-by: Mats Wichmann <mats@linux.com>