doc: change five functions to appear global-only
The five functions EnsureSConsVersion, EnsurePythonVersion, Exit,
GetLaunchDir, SConscriptChdir were listed as both global and
environment functions, but they do nothing in the context of an
environment, so marked in the xml as "global".  This only changes
the presentation in the manpage & userguide appendix, not the behavior.

Minor tweaks in the code around SConscriptChdir - actually use a
bool True/False instead of 0/1, and added a couple of type annotations.

Signed-off-by: Mats Wichmann <mats@linux.com>