Fix problem where Java inner classes cannot cache
Generated files contained a '$' in filename and this blew up subst.
Situation arose because of a need to fetch the FS entry of the
source for finding permissions. Now we use the permissions of the
cached target to decide whether to chmod to add write permission,
this avoids the need to call File() on the source.

Signed-off-by: Mats Wichmann <mats@linux.com>