Minor tex test fix
Syntax of selecting a glossary updated - apparently the late 2022.*
texlive releases have started warning and prompting for action on
this which appears to be what breaks the test.

Not directly related to this test, but to trying to get a reasonable
set of tests to run, added some pkgs to bin/scons_dev_master.py -
this has absolutely no effect on SCons itself.

Signed-off-by: Mats Wichmann <mats@linux.com>