Add a copyright header to test files
A lot of files that are not the main test scripts had no
license/copyright headers.  This change adds those.

In a small number of cases, this necessitaed a change to an expected
line number from a failure/exception message.

Along the way, some are lightly reformatted and some add a
DefaultEnvironment call.

Signed-off-by: Mats Wichmann <mats@linux.com>