pdb debugging now recognizes SConstruct/SConscript
Make it possible (once the project path has been added to sys.path)
to set/clear breakpoints using file paths that end in SConstruct and
SConscript - specializing pdb.Pdb to override the need for all paths
that are not absolute to end in .py.

Man: tried to address some surprise that pdb can't find the SConstruct
when you start up in debugger mode.

Signed-off-by: Mats Wichmann <mats@linux.com>