Fix dictifyCPPDEFINES handlong of macro strings
CPPDEFINES can contain name=value strings, either a single one, or one or
more in a sequence type.  After conversion (subsequent Append/Prepend to
CPPDEFINES), these will be stored as tuples, but it is possible to hit
cases where the type conversion has never been triggered. The C Scanner
has its own routine to process CPPDEFINES, and missed these cases,
which are now handled.

The testcases in issue 4193 which showed this problem are now included
in the C scanner unit tests, and the test for dictifyCPPDEFINES is
expanded to check these two forms.

Fixes #4193

Signed-off-by: Mats Wichmann <mats@linux.com>