Drop coverage from CI, don't add mingw
Experiment: coverage build broke because "codecov" package removed
from PyPI on 12 Apr 2023.  We'll need to re-craft a coverage build,
for now just don't set COVERAGE true for any appveyor build (leaving
the framework in place to copy to the new setup).

Experiment: the setup of mingw for the github action for experimental
features on Windows is failing for unknown reason ("error code 1"),
possibly sourceforge failure, possibly something else. But - we shouldn't
need to install mingw into an environment that is already provisioned
with mingw.  Try that.

Apparently it's technically a syntax error for setup.cfg lines (that are
not continuation lines) to be indented - had one tool gripe about this,
which was made happy by dedenting one line. Including that in this PR,
just because...

Signed-off-by: Mats Wichmann <mats@linux.com>