Add some cheap return and parameter annotations
Use: https://github.com/JelleZijlstra/autotyping
to add "safe" return annotations.

Where a parameter has a default value that is an obvious scalar type
(bool, int, str, etc.) add those annotations as well.

Also fixed two small bugs that popped up when sanity-checking with
mypy.  One in FortranCommon, where a return had been previously
annotated to be a tuple of Action, which should be ActionBase -
Action is the factory function, not the base class. The other was
a typo in the error raised in _add_cppdefines - the message was
formatted with the value of "define" which should have been "defines".

Signed-off-by: Mats Wichmann <mats@linux.com>