Fixup usage of new mock yacc tool
When the mock tool was revised to use argparse instead of getopt,
it stumbled across the problem that argparse apparently doesn't deal
with a single-letter option and its option-argument not having a
space between them until Python 3.8 - 3.6 and 3.7 failed the test.
Change to use the space in the test/mock tool.

Signed-off-by: Mats Wichmann <mats@linux.com>