Adjust help message about -H
As noted in issue #2356 (which this change does not complete/close),
the final line of help if the help message has been modified is
'Use scons -H for help about command-line options.'. This fails to
mention that -H prints only those options that are defined in SCons
itself, not options added via AddOption. Adjust the message to better
indicate this.

Signed-off-by: Mats Wichmann <mats@linux.com>