Remove script path in PlatformTests
For Python 3.12+, cpython module "posixpath" test-imports 'posix',
expecting it to fail on win32. However, for this one test program,
since it's in SCons/Platform, that directory as the "script path"
is in sys.path, and it finds our "posix" module, thus breaking things.
This removes that element from the path - it's not put back because
we don't actually need it here.

Fixes #4376

Signed-off-by: Mats Wichmann <mats@linux.com>