Add ability to save "local" help only
An additional keyword argument, "local_only", is now recorgnized
by Help(). If true, and "append" is True, then only project-defined
help messages are saved into the help text. That is, save help added
by calls to AddOption, but not SCons' own help, which is added
a different way.

Fixes #2356
Fixes #3686

Signed-off-by: Mats Wichmann <mats@linux.com>