Add LIBLITERAL to support gcc -l:filename
This is done in a general way. using a construction var, although
at the moment only the GNU linker is known to handle things this way.

Had to do something funky, or it won't work when os.pathsep and
LIBLITRAL are the same, as they are on Linux (i.e. ':'). That's because
SCons.PathList.PathList is called and it does:

class _Pathlist:
    def __init__(self, pathlist) -> None:
        if SCons.Util.is_String(pathlist):
            pathlist = pathlist.split(os.pathsep)

Splitting has the effect of turning ":libm.a" into ((0, ''), (0, 'libm.a')]
That is, the ':' is lost as part of the library specifier - so need to
try to avoid that.

Fixes #3951

Signed-off-by: Mats Wichmann <mats@linux.com>