Handle json decode error in mavc cache
A bit of a revamp of read_script_env_cache and write_script_env_cache.
The significant change is that on read, a JSDONDecodeError is
detected, and the cache file removed if so - we're guessing a
write race corrupted the file.

If this works, will update with more "PR polish", else will withdraw.

Signed-off-by: Mats Wichmann <mats@linux.com>