msvc config cache: add locking on cachefile
Following on from #4402, which tries to recover from races that might
either cause a corrupt msvc cache file, or an incomplete read happening
while a write is still in process, add a simple-minded locking protocol to
try to prevent the problem in the first place.  For writing, a lockfile
is created with exclusive access. For reading, the same is done but
immediately released: we only wanted to wait that it's not currently
locked, we don't need to keep it locked at this point.

This is addressing what's mainly an issue for testing, when there can
be many concurrent SCons instance each running a test - we don't think
this is likely in normal developer usage.

Signed-off-by: Mats Wichmann <mats@linux.com>