Tweak Util module
Move UtilTests.py from the top level to the Util package directory,
for consistency with the other packages with unittests.

Renamed Util/types.py -> Util/sctypes.py.  types is the name of a stdlib
module and it's a bad idea to duplicate it, even though in this case it's
"legal" since the file was not at the top level.  (Moving UtilTests.py
actually made this a real problem)

Class Selector is no longer an OrderedDict, it just inherits from dict
as ordering is now preserved and we never used any extra features of
OrderedDict.

Fix API doc build - was missing a good bit of Util since it was split
into a package.

Moved the import-loop warning to the top of __init__.py so it will be
more visible.

Fiddly linting, doc-stringing, etc.  Super-fiddly: pylint flags foo, bar
and baz as prohibited variable/function/method names.  Actually changed
these in UtilTests.py.

SCons.Errors has been a source of import loops because it imports Util.
Now Util is split, directly import from the Util.sctypes submodule the
two things Errors needs - this may reduce the chance of import problems.

Signed-off-by: Mats Wichmann <mats@linux.com>