Bump dependency versions  [skip appveyor]
Sphinx ceiling version bumped to 6.x. While it *works* with current
Sphinx 7.x, sphinx-book-theme is actually pinned to <7, so to avoid
install complaints, leave Sphinx down-rev for now.

lxml version pin bumped, and the fence for not installing on
Python 3.12 is changed to 3.13 - PyPI packages for 3.12 are up.

Update doc/generated/variables.mod as it contains new construction
variables from the 4.6 development: all the generated files will be
updated at release time, but this allows a build *without*
regenerating to not fail.

Signed-off-by: Mats Wichmann <mats@linux.com>