Further mod on msvc config cache locking
In the first iteration, we took a read lock in the read case.
A fail was observed where the reader took a json decode error,
and then tried to remove the cachefile in case it's really corrupt.
That failed as the file was "busy" - on Windows removing is
essentially writing the file - so take a write lock here too.

Signed-off-by: Mats Wichmann <mats@linux.com>