Fix ninja warnings/exceptions
Try to normalize errors/warnings.  One warning which was just a bare
class instantition (thus, a no-op) now warns; two failures which used
to raise SConsWarning now raise ImportError since that's what they are
(has no visible effect to the user); another now raises UserError,
as it is raised for an error in SConscripts.

Fixes #4036

Signed-off-by: Mats Wichmann <mats@linux.com>