Remove incorrect dependencies from `setup.cfg`
Remove the `install_requires` on `setuptools`.  This key is used to
specify packages that are needed at runtime.  SCons nowhere in its code
does import `setuptools` or `pkg_resources`.

Remove the `setup_requires` on `build`.  `build` is a frontend package
and a detail of how the build is invoked, while `setup_requires` are
used to specify backend dependencies (i.e. packages that are installed
after `build` is invoked).

Remove the `setup_requires` on `setuptools`.  It is a key specific
to setuptools, so for it to be interpreted `setuptools` need to be
installed already.  The actual backend dependency on `setuptools`
is specified in `pyproject.toml`, so the dependency is entirely
redundant.